refactor: update notificationType filename
